pyproject.toml
[tool.poetry]
name = "sir-robin"
version = "0.1.0"
description = ""
authors = ["Python Discord <info@pythondiscord.com>"]

[tool.poetry.dependencies]
python = "3.9.*"
"discord.py" = {url = "https://github.com/Rapptz/discord.py/archive/beafaa8a8b8006357c2a1ca6802ed80e000f4cda.zip"}
bot-core = {url = "https://github.com/python-discord/bot-core/archive/refs/tags/v4.0.0.zip"}
aiodns = "^3.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-annotations = "^2.7.0"
flake8-bugbear = "^22.1.11"
flake8-docstrings = "^1.6.0"
flake8-isort = "^4.1.1"
flake8-string-format = "^0.3.0"
flake8-tidy-imports = "^4.6.0"
flake8-todo = "^0.7"
pep8-naming = "^0.12.1"
pre-commit = "^2.17.0"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
hypothesis = "^6.14.0"
python-dotenv = "^0.19.2"
taskipy = "^1.10.1"

[tool.taskipy.tasks]
start = "python -m bot"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
